requirements.txt
alabaster==1.0.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
babel==2.17.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
docutils==0.21.2
fastapi==0.115.12
h11==0.14.0
idna==3.10
imagesize==1.4.1
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
psycopg2-binary==2.9.10
pydantic==2.11.2
pydantic_core==2.33.1
Pygments==2.19.1
python-dotenv==1.1.0
requests==2.32.3
roman-numerals-py==3.1.0
sniffio==1.3.1
snowballstemmer==2.2.0
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.4.0
uvicorn==0.34.0
watchfiles==1.0.5
websockets==15.0.1
